refactor: move peerconnections
